Move silenced_es_logger to azul.logging
